+++ Data/digit_load.hpp
@@ -45,7 +45,7 @@ T swap_endian(T u)
 
     source.u = u;
 
-    for (size_t k = 0; k < sizeof(T); k++)
+    for (int k = 0; k < sizeof(T); k++)
         dest.u8[k] = source.u8[sizeof(T) - k - 1];
 
     return dest.u;
